==> notebooks/1-API_ASOS_extract2.ipynb <==
# EXTRACT
# Extracting Women clothes from ASOS API (2)


(2) Extract more products similar to t-shirt category

### 0. Import modules

import requests
import time
import pandas as pd
import pickle

from logs.PASSES import RAPID ## my TOKEN

len(RAPID)

## RapidAPI - Asos

What is RapidAPI?
RapidAPI is an API Marketplace for developers to find, connect, and manage their API connections. 

https://rapidapi.com/apidojo/api/zappos1

headers = {
    "X-RapidAPI-Key": RAPID,
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

### Get countries available to get ASOS data from

Extract the relevant info for **SPAIN**

url = "https://asos2.p.rapidapi.com/countries/list"

response = requests.get(url, headers=headers)

countries_res=response.json()

ES_index = [index for index, countries in enumerate(countries_res) if countries.get("country") == "ES"]
nombre_pais=countries_res[ES_index[0]]["name"]
pais=countries_res[ES_index[0]]["country"]
idioma=countries_res[ES_index[0]]["languages"][0]["language"]
moneda=countries_res[ES_index[0]]["currencies"][0]["currency"]

#display(nombre_pais, pais, idioma, moneda)

### 5 main brand categories
keep `WW_FP A-Z Women's Brands`

url = "https://asos2.p.rapidapi.com/categories/list"
querystring = {"country":"pais", "lang":"idioma"}
response = requests.get(url, headers=headers, params=querystring)
brands_res=response.json()

[print(R["alias"],R["content"]["title"]) for R in brands_res["brands"]] 
main_cats=brands_res["brands"][2] ## WW_FP A-Z Women's Brands

### Brand ids in `WW_FP A-Z Women's Brands`

[print(R["content"]["title"], R["link"]["categoryId"]) for R in main_cats["children"]]

#Roxy 7274
#Quiksilver 28829
#Lacoste 7294

### Get data directly from category id

url = "https://asos2.p.rapidapi.com/categories/list"

querystring = {"country":"pais", "lang":"idioma"}

response = requests.get(url, headers=headers, params=querystring)

res=response.json()

res["navigation"][0] #men
res["navigation"][1] #women
resWomen=res["navigation"][1]

### Create a list containing the type of items and their category id

category_list=[]
for RESW in resWomen["children"]:
   # display(RESW["children"])
    for i in range(len(RESW["children"])):
        RESW2=RESW["children"][i]["children"]
        for j in RESW2:
            RESW3=j["children"]
            for n in RESW3:
                #display(n["link"]["categoryId"])
                #display(n["link"]["webUrl"])
                category_list.append(n["link"]["webUrl"])

category_list2=[CAT.split("cid=") for CAT in category_list]
category_list2=[CAT for CAT in category_list2 if len(CAT)==2]


### CategoryId of Woman T-shirts is # 20581

#find_category="t-shirt"
#find_category="top"
find_category="shirt"


category_listx=[]
for CAT in category_list2:
    if(find_category in CAT[0]):
        category_listx.append(CAT)
        
category_listx[:3]

### "shirts" category

#find_category="t-shirt"
#find_category="top"
find_category="shirt"

category_listx=[]
for CAT in category_list2:
    if(find_category in CAT[0]):
        category_listx.append(CAT)
        
category_listx

categories=[VALS[1] for VALS in category_listx]
categories=list(set(categories))
categories_shirt=categories

### "tops" category

#find_category="t-shirt"
find_category="top"
#find_category="shirt"

category_listx=[]
for CAT in category_list2:
    if(find_category in CAT[0]):
        category_listx.append(CAT)

category_listx=[VALS for VALS in category_listx if "brands" not in  VALS[0]]
category_listx=[VALS[1] for VALS in category_listx]
categories=[VALS.split("&")[0] for VALS in category_listx]
categories_top=categories
categories_top

find_category="t-shirt"
#find_category="top"
#find_category="shirt"

category_listx=[]
for CAT in category_list2:
    if(find_category in CAT[0]):
        category_listx.append(CAT)
category_listx     

categories=[VALS[1] for VALS in category_listx]
categories=list(set(categories))
categories=[VALS.split("&")[0] for VALS in categories]
categories_top=categories

### "dresses" category

#find_category="t-shirt"
find_category="dresses"
#find_category="shirt"

category_listx=[]
for CAT in category_list2:
    if(find_category in CAT[0]):
        category_listx.append(CAT)

category_listx=[VALS for VALS in category_listx if "brands" not in  VALS[0]]
category_listx=[VALS[1] for VALS in category_listx]
categories=[VALS.split("&")[0] for VALS in category_listx]
categories_dress=categories


### Iterate the search to get all the products available

Asos has a limit of 48 products by request.

Change the category Id to extract the products of each category


url = "https://asos2.p.rapidapi.com/products/v2/list"
headers = {
    "X-RapidAPI-Key": RAPID,
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

# categories_dress
# "20581" #only t-shirt
#categories_top
#categories_shirt


category_id = categories_dress # "20581"
limit = 48
offset = 0
results = []

while True:
    querystring = {
        "categoryId": categories_dress, #"20581",
        "lang":{idioma},"currency":{moneda},"store":{pais},"country":{pais},
        "limit": str(limit),
        "offset": str(offset),
        "sort": "freshness"
    }

    response = requests.get(url, headers=headers, params=querystring)

    if response.status_code == 200:
        data = response.json()
        results.extend(data['products'])  # Append the products from the current page to results
        offset += limit  # Increment the offset to get the next page
        if len(data['products']) < limit:
            break  # Break the loop if the current page has fewer than the limit
    else:
        print(f"Error: {response.status_code}")
        break  # Exit the loop on error



### Parse json to DataFrame

thedf=[[row["brandName"], row["name"],  row["price"]["current"]["text"], row["colour"]] for row in results]
tets_df=pd.DataFrame(thedf,columns=["brand", "description", "price", "colour"])
tets_df[:2]

### Save the data using pickle

The Lines are commented to avoid overwritting

# asos_womendress_data=results #only t-shirts
## asos_womentshirt_data=results #only t-shirts
## asos_womenshirt2_data=results # all "shirts"
# asos_womentop_data=results # all "shirts"

### Save the object to a file using pickle
#with open('asos_womendress_data.pkl', 'wb') as file:
#    pickle.dump(asos_womendress_data, file)

### Save the object to a file using pickle
#with open('asos_womentop_data.pkl', 'wb') as file:
#    pickle.dump(asos_womentop_data, file)

### Save the object to a file using pickle
#with open('asos_womenshirt2_data.pkl', 'wb') as file:
#    pickle.dump(asos_womenshirt2_data, file)
    
### Save the object to a file using pickle
#with open('asos_womentshirt_data.pkl', 'wb') as file:
#    pickle.dump(asos_womentshirt_data, file)

with open('../../data/raw/asos_womentshirt_data.pkl', 'rb') as file:
    asos_womentshirt_data = pickle.load(file)

asos_raw=asos_womenshirt2_data

thedf2=[[row["brandName"], row["name"],  row["price"]["current"]["text"], row["colour"]] for row in asos_raw]
color_df2=pd.DataFrame(thedf2,columns=["brand", "description", "price", "colour"])
color_df2.tail()
