==> notebooks/2-API_ASOS_transform2.ipynb <==
# TRANSFORM
# Cleaning raw data from Women clothes from ASOS API (2)

Data was accessed through an API on 2023-11-04 (t-shirt category) and 2023-11-09 (the rest of categories)<br/>
stored as a pickle file in `../../data/raw/ *** .pk`

The raw datasets are json files with a very complex structure <br/>

Now the data must be cleaned and transform to comply with the DDBB standards.

--------

Raw `asos_womentshirt_data` dataset has the following issues:


* 1. Extract only relevant data **(brand, description, price, colour)**
    
* 2. ASOS dataset **486 names** for colours
    * Normalize the color names using `mlg.namvector_clean`
    * Simplify colour names according to the colors used by Amazon.es
    * Use `../../data/raw/color_simplification.pkl` - A chatGPT generated dictionary to simplifify the colors returned by ASOS
    * Transform rows with multiple color values transform to "multicolor"

* 3. Clean `price` column
* 4. Normalize brand names

--------

Save cleaned data in `../../data/clean/asos_womentshirt_data.pk`

### 0. Import the modules

import requests
import time
import pandas as pd
import pickle
import numpy as np
import warnings
warnings.filterwarnings('ignore') # ignorar warnings
from src import dataanalysis_fun1 as mlg  #Import my module

Reload my module if neccessary

import importlib
from src import dataanalysis_fun1 as mlg # Import the module
#importlib.reload(mlg)  # Reload the module

# Suppress warning when reloading the module
with warnings.catch_warnings():
    warnings.simplefilter("ignore") 
    importlib.reload(mlg)  # Reload the module

## 1. Load the data using pickle

LOAD ADDITIONAL DATA FROM ASOS (DRESSES, TOPS AND SWEATSHIRTS)

# Now, let's load the object from the file
with open('../../data/raw/asos_womentshirt_data.pkl', 'rb') as file:
    asos_womentshirt_data = pickle.load(file)

# Now, let's load the object from the file
with open('../../data/raw/asos_womentop_data.pkl', 'rb') as file:
    asos_womentop_data = pickle.load(file)

# Now, let's load the object from the file
with open('../../data/raw/asos_womentop_data.pkl', 'rb') as file:
    asos_womentop_data = pickle.load(file)

# Now, let's load the object from the file
with open('../../data/raw/asos_womendress_data.pkl', 'rb') as file:
    asos_womendress_data = pickle.load(file)


asos_raw1=asos_womentshirt_data
asos_raw2=asos_womentop_data
asos_raw3=asos_womentop_data
asos_raw4=asos_womendress_data

## 2. Transform (clean the data)

#### Clean and extract relevant text

The ultimate goal is, for each element, to keep only the `brand, product type, colour and price €` 

## 2.1 Join the datasets and parse them into DataFrame

asos_raw_list=[asos_raw1,asos_raw2, asos_raw3, asos_raw4]

color_df_list=[]
for ASOS in asos_raw_list:
    thedf1=[[row["brandName"], row["name"],  row["price"]["current"]["text"], row["colour"]] for row in ASOS]
    color_df=pd.DataFrame(thedf1,columns=["brand", "description", "price", "colour"])
    color_df_list.append(color_df)
    
ASOS_df = pd.concat(color_df_list, ignore_index=True)

# delete duplicated rows 

ASOS_df1=ASOS_df[~ASOS_df["description"].duplicated()]
ASOS_df1.reset_index()

#### ASOS dataset has a huge number of ways to clasify the products' colours

len(ASOS_df1["colour"].unique())

## 2.2 Normalize the color names using `mlg.namvector_clean`

ASOS_df1=ASOS_df1.copy()
ASOS_df1["colour"]=mlg.namvector_clean(ASOS_df1["colour"])
ASOS_df1["colour"]=[col.replace("_", " ") for col in ASOS_df1["colour"]]

### 2.3 Simplify colour names according to the color categories used by Amazon.es

amz_colors=['negro', 'gris', 'blanco', 'marron', 'beis', 'rojo', 'rosa',
       'naranja', 'amarillo', 'marfil', 'verde', 'turquesa', 'azul',
       'morado', 'dorado']

english_colors = ['black', 'gray', 'white', 'brown', 'beige', 'red', 'pink', 'orange', 'yellow', 'ivory', 'green', 'turquoise', 'blue', 'purple', 'gold']
amz_colors_dict = dict(zip(english_colors, amz_colors))

print(amz_colors_dict)

### 2.3 a - 16 SPANISH NAMES

ASOS_df2=ASOS_df1.copy()
ASOS_df2["colour_simp"]=""

for color in amz_colors:
    '''
    check in original color nomenclature if a simplest color is present
    some nomenclatures contain multiple words, i need doble comprehension to eval all the words by cell
    '''
    matches = [any(word in color for word in cell.split())
               for cell in ASOS_df2["colour"].values]
    
    ASOS_df2.loc[matches, "colour_simp"] = color

# ASOS_df2["colour_simp"].value_counts()

### 2.3 b -16 ENGLISH NAMES


ASOS_df2.head()

for dict_key, dict_value in amz_colors_dict.items(): 
    matches = [cell in dict_key for cell in ASOS_df2["colour"].values]
    
    '''
    check in original color nomenclature if a simplest color is present
    some nomenclatures contain multiple words, i need doble comprehension to eval all the words by cell
    '''
    matches = [any(word in dict_key for word in cell.split())
               for cell in ASOS_df2["colour"].values]
    
    ASOS_df2.loc[matches, "colour_simp"] = dict_value


# ASOS_df2["colour_simp"].value_counts()

### 2.3 c - `collapsed_dict2.pkl` -chatGPT generated dictionary
Load the object from the file -chatGPT generated dictionary to simplifify the colors returned by ASOS

with open('../../data/raw/collapsed_dict2.pkl', 'rb') as file:
    collapsed_dict2 = pickle.load(file)

collapsed_dict2.keys()

def color_dict_simplify(x):
    x = x.lower()
    theval=""
    for KKK, VVV in collapsed_dict2.items():
        for VAL in VVV:
            if VAL in x:
                theval = KKK
            
    return theval

ASOS_df3=ASOS_df2.copy()

ASOS_df3['colour_simp2'] = ASOS_df3['colour'].apply(color_dict_simplify)
#ASOS_df3['colour_simp2'].value_counts()

### 2.3 d- Unify colors in `colour_simp` and keep only one column

def update_colour_simp(row):
    if row['colour_simp'] == "":
        return row['colour_simp2']
    else:
        return row['colour_simp']

ASOS_df3['colour_simp'] = ASOS_df3.apply(update_colour_simp, axis=1)


#### There still a crazy ASOS names 

We have still 269 rows without simple nomenclature. <br/>
I could Use fuzzywuzzy to clean it but -for the moment-, I prefer just to drop them

crazy_ASOS=ASOS_df3[ASOS_df3["colour_simp"]==""]

print(len(crazy_ASOS))
display(crazy_ASOS.head(5))

### 2.3 e- Drop rows with crazy names and keep only simplified color

ASOS_df4=ASOS_df3.copy()
ASOS_df4 = ASOS_df4.loc[:, ['brand', 'description', 'price', "colour_simp"]]
ASOS_df4 = ASOS_df4.rename(columns={'colour_simp': 'colour'})


ASOS_df4.colour =[row.lstrip() for row in ASOS_df4.colour]
ASOS_df4=ASOS_df4[ASOS_df4["colour"]!=""]
ASOS_df4.sample(10)

#### Simplify ` turquesa` for `azul` and `marfil` for `beis`

ASOS_df4["colour"][ASOS_df4["colour"]=="turquesa"]="azul"
ASOS_df4["colour"][ASOS_df4["colour"]=="marfil"]="beis"

ASOS_df4["colour"].value_counts()

## 2.4 Clean `price` column

* Take out € symbol
* Replace , for .
* Transform to float

ASOS_df4[:5]

ASOS_df4["price"]=[VAL.replace(",", ".") for VAL in ASOS_df4["price"]]
ASOS_df4["price"]=[VAL.replace("€", "") for VAL in ASOS_df4["price"]]
ASOS_df4["price"]=[float(VAL) for VAL in ASOS_df4["price"]]

ASOS_df4[:5]

## 2.5 Normalize `brand` names

ASOS_df5=ASOS_df4.copy()
ASOS_df5.brand = [row.replace(".", "") for row in ASOS_df5.brand]

display(ASOS_df5.groupby('colour')['price'].mean())
display(ASOS_df5.groupby('brand')['price'].mean())

## 3. Save cleaned data

#asos_womenupperclothes_clean=ASOS_df5.copy()
#asos_womenupperclothes_clean.to_csv('../../data/clean/asos_womenupperclothes_clean.csv', index=False)
